Fix for KOIOS_API_HEADERS quoting issue
Fixes a quoting issue in the KOIOS_API_HEADERS array. (redundantly quoting all elements as a single string).

This led to the the koios api call being unauthenticated and so hitting the rate limit of 5k request per day which is the 'public' (unauthenticated) tier, even though a valid KOIOS_API_TOKEN was defined.